Emergency requisites
This commit completes the functional requisites from the `Emergency`
package.

Also it modifies the class diagram, adding the `EmergencyReport` class.

It also modifies some use cases, mainly adding `Log in into the system`, and
removing `See emergency comments` (which is now integrated in `See
emergency details`) and `Add reporter` (which is integrated in `Report
emergency`).